REFACTOR move all logic to separate files
